Corrected cdb for Clear Effector
I have corrected the .cdb file for Clear Effector. The string str2 was missing, meaning that when Clear Effector was used as a Synchro Material, the tooltip (the little box that shows when you hover your mouse over a card) for the Synchro Summoned monster showed "*???" instead of "*Cannot be destroyed by card effects". I have added the string to the above cdb so that it will show.